+++ test-module/src/ref_cell.rs
@@ -28,6 +28,6 @@ fn inc(x: &mut i64) -> Result<i64> {
 #[defun]
 #[allow(clippy::trivially_copy_pass_by_ref)] // TODO: Test with sth else not i64.
 fn unwrap_and_call(_: &i64, lambda: Value<'_>) -> Result<()> {
-    lambda.env.call(lambda, [])?;
+    lambda.call([])?;
     Ok(())
 }